requirements.txt
requests
rich
fake-useragent
fastapi
uvicorn
python-dotenv
httpx
pydantic
typing-extensions